chore(DateTimeUtils): move tests
