+++ packages/fluentui/react-northstar/test/specs/utils/date-time-utils/dateGrid/getBoundedDateRange-test.ts
@@ -1,4 +1,4 @@
-import * as DateGrid from './getBoundedDateRange';
+import * as DateGrid from '../../../../../src/utils/date-time-utilities/dateGrid/getBoundedDateRange';
 
 enum Months {
   Jan = 0,
